et tu _lu_solve

+++ gwcs/wcs/_utils.py
@@ -45,7 +45,8 @@ def _poly_fit_lu(xin, yin, xout, yout, degree, coord_pow=None):
 
     nterms = len(powers)
 
-    flt_type = np.longdouble
+    # lu_factor and _lu_solve deprecated float128 support in scipy 1.18.0
+    flt_type = np.float64
 
     # allocate array for the coefficients of the system of equations (a*x=b):
     a = np.empty((nterms, nterms), dtype=flt_type)
@@ -96,7 +97,7 @@ def _poly_fit_lu(xin, yin, xout, yout, degree, coord_pow=None):
     with warnings.catch_warnings(record=True):
         warnings.simplefilter("error", category=linalg.LinAlgWarning)
         try:
-            lu_piv = linalg.lu_factor(a.astype(np.float64))
+            lu_piv = linalg.lu_factor(a)
             poly_coeff_x = linalg.lu_solve(lu_piv, bx).astype(float)
             poly_coeff_y = linalg.lu_solve(lu_piv, by).astype(float)
         except (ValueError, linalg.LinAlgWarning, np.linalg.LinAlgError) as e:
